- Merged develop - General fixes and alterations
