+++ pom.xml
@@ -84,6 +84,7 @@
     <groupId>de.flapdoodle.embed</groupId>
     <artifactId>de.flapdoodle.embed.mongo</artifactId>
     <scope>test</scope>
+    <version>4.16.2</version>
   </dependency>
 	</dependencies>
 	<build>
