--- src/main/java/com/bookflux/aplication/service/BookCollectionApiService.java
@@ -1,9 +1,0 @@
-package com.bookflux.aplication.service;
-
-import com.bookflux.dto.GoogleBooksResponseDto;
-import com.google.api.client.googleapis.services.GoogleClientRequestInitializer;
-
-public interface BookCollectionApiService {
-
-    public GoogleBooksResponseDto searchBook(String query);
-}
