+++ src/main/java/com/bookflux/integration/mapper/BookMapper.java
@@ -1,0 +1,52 @@
+package com.bookflux.integration.mapper;
+
+
+import com.bookflux.dto.GoogleBooksResponseDto;
+import com.bookflux.dto.ImagelinksDto;
+import com.bookflux.enums.MaturityRating;
+import com.bookflux.repository.collection.BookCollection;
+
+public final class BookMapper {
+
+  private BookMapper() {
+
+  }
+
+  public static BookCollection toDomain(GoogleBooksResponseDto responseDto) {
+    if (responseDto == null || responseDto.getItems() == null || responseDto.getItems().isEmpty()) {
+      return null;
+    }
+
+    GoogleBooksResponseDto.VolumeInfo info = responseDto.getItems().getFirst().getVolumeInfo();
+
+    return BookCollection.builder()
+        .id(null)
+        .title(info.getTitle())
+        .authors(info.getAuthors())
+        .publisher(info.getPublisher())
+        .publishedDate(info.getPublishedDate())
+        .description(info.getDescription())
+        .pageCount(info.getPageCount())
+        .categories(info.getCategories())
+        .maturityRating(convertMaturity(info.getMaturityRating()))
+        .averageRating(info.getAverageRating())
+        .imageLinks(mapImageLinks(info))
+        .language(info.getLanguage())
+        .build();
+  }
+
+  private static ImagelinksDto mapImageLinks(GoogleBooksResponseDto.VolumeInfo info) {
+    if (info.getImageLinks() instanceof ImagelinksDto dto) {
+      return dto;
+    }
+    return null;
+  }
+
+  private static MaturityRating convertMaturity(String value) {
+    try {
+      return MaturityRating.valueOf(value.toUpperCase());
+    } catch (Exception e) {
+      return null;
+    }
+  }
+}
