+++ src/main/java/com/bookflux/integration/service/BookCollectionApiService.java
@@ -1,0 +1,8 @@
+package com.bookflux.integration.service;
+
+import com.bookflux.dto.GoogleBooksResponseDto;
+
+public interface BookCollectionApiService {
+
+    public GoogleBooksResponseDto searchBook(String query);
+}
